requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
distro==1.9.0
Elixir==0.7.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
JPype1==1.5.0
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.23.6
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pbr==6.0.0
pdfminer.six==20231228
pefile==2023.2.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.0
tabula-py==2.9.0
Tempita==0.5.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
xlrd==0.7.1
xlwt==0.7.2
